+++ tests/common/sql_test_case.py
@@ -21,6 +21,7 @@ from sodasql.scan.db import sql_update, sql_updates
 from sodasql.scan.dialect import Dialect
 from sodasql.scan.dialect_parser import DialectParser
 from sodasql.scan.env_vars import EnvVars
+from sodasql.scan.metric import Metric
 from sodasql.scan.scan_result import ScanResult
 from sodasql.scan.scan_yml_parser import KEY_TABLE_NAME, ScanYmlParser
 from sodasql.scan.sql_metric_yml_parser import SqlMetricYmlParser
@@ -51,8 +52,8 @@ class SqlTestCase(TestCase):
     warehouse_fixture_cache_by_target = {}
     warehouses_close_enabled = True
 
-    def __init__(self, methodName: str = ...) -> None:
-        super().__init__(methodName)
+    def __init__(self, method_name: str = ...) -> None:
+        super().__init__(method_name)
         self.warehouse: Optional[Warehouse] = None
         # self.target controls the warehouse on which the test is executed
         # It is initialized in method setup_get_warehouse
@@ -152,6 +153,57 @@ class SqlTestCase(TestCase):
                                           sql_metrics=sql_metrics,
                                           variables=variables)
         return scan.execute()
+
+    def execute_metric(self, warehouse: Warehouse, metric: dict, scan_dict: dict = None):
+        dialect = warehouse.dialect
+        if not scan_dict:
+            scan_dict = {}
+        if KEY_TABLE_NAME not in scan_dict:
+            scan_dict[KEY_TABLE_NAME] = self.default_test_table_name
+        scan_configuration_parser = ScanYmlParser(scan_dict, 'Test scan')
+        scan_configuration_parser.assert_no_warnings_or_errors()
+        scan = warehouse.create_scan(scan_configuration_parser.scan_yml)
+        scan.execute()
+
+        fields: List[str] = []
+        group_by_column_names: List[str] = metric.get('groupBy')
+        if group_by_column_names:
+            for group_by_column in group_by_column_names:
+                fields.append(dialect.qualify_column_name(group_by_column))
+
+        column_name: str = metric.get('columnName')
+        qualified_column_name = dialect.qualify_column_name(column_name)
+
+        metric_type = metric['type']
+        if metric_type == Metric.ROW_COUNT:
+            fields.append('COUNT(*)')
+        if metric_type == Metric.MIN:
+            fields.append(f'MIN({qualified_column_name})')
+        elif metric_type == Metric.MAX:
+            fields.append(f'MAX({qualified_column_name})')
+        elif metric_type == Metric.SUM:
+            fields.append(f'SUM({qualified_column_name})')
+
+        sql = 'SELECT \n  ' + ',\n  '.join(fields) + ' \n' \
+              'FROM ' + scan.qualified_table_name
+
+        where_clauses = []
+
+        metric_filter = metric.get('filter')
+        if metric_filter:
+            where_clauses.append(dialect.sql_expression(metric_filter))
+
+        scan_column: ScanColumn = scan.scan_columns.get(column_name)
+        if scan_column and scan_column.non_missing_and_valid_condition:
+            where_clauses.append(scan_column.non_missing_and_valid_condition)
+
+        if where_clauses:
+            sql += '\nWHERE ' + '\n      AND '.join(where_clauses)
+
+        if group_by_column_names:
+            sql += '\nGROUP BY ' + ', '.join(group_by_column_names)
+
+        return warehouse.sql_fetchall(sql)
 
     def assertMeasurements(self, scan_result, column: str, expected_metrics_present):
         metrics_present = [measurement.metric for measurement in scan_result.measurements
